refactor: 그룹챗 수정 삭제로직 분리
